Split memory.h into definition and implementation headers (lib)
